requirements.txt
click==7.1.2
colorama==0.4.3
deprecation==2.1.0
eyeD3==0.9.5
ffmpeg-python==0.2.0
filetype==1.0.7
future==0.18.2
logzero==1.5.0
packaging==20.3
pyparsing==2.4.7
pytube3==9.6.4
six==1.14.0
typing-extensions==3.7.4.2
